refactor(voice): add lightweight response guard registry
